Initial Cleanup to remove unused files from this repo Rework README.md, RELEASES.md Remove extension sub-folder Update Build steps to handle removal of extension folder
