+++ .github/workflows/ci.yml
@@ -5,9 +5,9 @@ name: Node.js CI
 
 on:
   push:
-    branches: [ master, beta, Release2.7 ]
+    branches: [ main ]
   pull_request:
-    branches: [ master, beta, Release2.7 ]
+    branches: [ main ]
 
 jobs:
   build:
