+++ .github/workflows/codeql-analysis.yml
@@ -13,10 +13,10 @@ name: "CodeQL"
 
 on:
   push:
-    branches: [ master, beta, Release2.7 ]
+    branches: [ main ]
   pull_request:
     # The branches below must be a subset of the branches above
-    branches: [ master, beta, Release2.7 ]
+    branches: [ main ]
   schedule:
     - cron: '17 15 * * 2'
