+++ applicationinsights-react-native/Tests/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "compilerOptions": {
+        "sourceMap": true,
+        "inlineSources": true,
+        "noImplicitAny": false,
+        "module": "amd",
+        "moduleResolution": "Node",
+        "target": "es5",
+        "alwaysStrict": true,
+        "declaration": true,
+        "outFile": "./Selenium/reactnativeplugin.tests.js"
+      },
+      "include": [
+      ]
+}
