+++ applicationinsights-react-native/src/Interfaces/INativeDevice.ts
@@ -1,0 +1,16 @@
+export interface INativeDevice {
+    /**
+     * Device type, e.g. Handset, Tablet, Tv
+     */
+    deviceClass?: string;
+
+    /**
+     * Unique installation ID
+     */
+    id?: string;
+
+    /**
+     * The device model: iPhone XS Max, Galaxy S10, etc
+     */
+    model?: string;
+}
