+++ applicationinsights-react-native/src/Interfaces/IReactNativePluginConfig.ts
@@ -1,0 +1,4 @@
+export interface IReactNativePluginConfig {
+    disableDeviceCollection?: boolean;
+    disableExceptionCollection?: boolean;
+}
