+++ applicationinsights-react-native/src/Interfaces/index.ts
@@ -1,0 +1,4 @@
+import { INativeDevice } from "./INativeDevice";
+import { IReactNativePluginConfig } from "./IReactNativePluginConfig";
+
+export { INativeDevice, IReactNativePluginConfig };
