+++ applicationinsights-react-native/src/index.ts
@@ -1,0 +1,4 @@
+import {ReactNativePlugin} from "./ReactNativePlugin";
+import { INativeDevice, IReactNativePluginConfig } from "./Interfaces";
+
+export { ReactNativePlugin, INativeDevice, IReactNativePluginConfig };
