+++ applicationinsights-react-native/tslint.json
@@ -1,0 +1,5 @@
+{
+    "extends": [
+        "../../tslint-base.json"
+    ]
+}
