--- extensions/applicationinsights-angularplugin-js/README.md
@@ -1,25 +1,0 @@
-# Microsoft Application Insights JavaScript SDK - Angular Plugin (&lt; 2.5.10)
-
-*******************
-
-:exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation:
-
-# The current development repo location of the [Angular Plugin Repo has moved here](https://github.com/microsoft/applicationinsights-angularplugin-js) - (for v2.5.11 or greater) 
-
-## Please refer to the new repo for usage instructions, and open issues there.
-
-### This repo folder is no longer compiled and tested as part of the PR and merge validation, thus it WILL go stale
-
-:exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation: :exclamation:
-
-*******************
-
-### Where is the source code for the previous version
-
-For historical purposes the previous code for the angular plugin has been archived in the [angularplugin-legacy branch to review all previous code and documentation](https://github.com/microsoft/ApplicationInsights-JS/tree/angularplugin-legacy/extensions/applicationinsights-angularplugin-js) branch.
-
-No further changes are scheduled or expected, any and all changes will occur in the [new repo](https://github.com/microsoft/applicationinsights-angularplugin-js) only.
-
-### Links
-
-- [v2.5.10 Readme](https://github.com/microsoft/ApplicationInsights-JS/blob/angularplugin-legacy/extensions/applicationinsights-angularplugin-js/README.md)
