--- extensions/applicationinsights-react-native/.npmignore
@@ -1,15 +1,0 @@
-# NPM Ignore
-
-# ignore everything
-*
-
-# ... but these files
-!package.json
-!tsconfig.json
-!/README.md
-!/LICENSE
-!dist-esm/**
-!dist/**
-!browser/**
-!types/**
-!src/**
