--- extensions/applicationinsights-react-native/Tests/UnitTests.html
@@ -1,63 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="Cache-control" content="no-Cache" />
-    <title>Tests for Application Insights JavaScript API</title>
-    <link rel="stylesheet" href="https://code.jquery.com/qunit/qunit-1.23.1.css">
-    <script src="http://sinonjs.org/releases/sinon-2.3.8.js"></script>
-    <script src="http://cdnjs.cloudflare.com/ajax/libs/require.js/2.2.0/require.js"></script>
-    <script src="../../../common/Tests/Selenium/ModuleLoader.js"></script>
-    <script src="../../../common/Tests/Selenium/SimpleSyncPromise.js"></script>
-
-    <script>
-        var modules = new ModuleLoader({
-            baseUrl: '../',
-            paths: {
-                qunit: "../../common/Tests/External/qunit-1.23.1",
-                "react-native": "./Tests/External/DummyReactNative"
-            }
-        });
-
-        // Load qunit here instead of with tests, otherwise will not work
-        modules.add("qunit");
-
-        // Load and define the app insights test framework module
-        modules.add("@microsoft/ai-test-framework", "./node_modules/@microsoft/ai-test-framework/dist/ai-test-framework");
-
-        // Load and define the app insights Shims module
-        modules.add("@microsoft/applicationinsights-shims", "./node_modules/@microsoft/applicationinsights-shims/browser/applicationinsights-shims");
-
-        // Load DynamicProto
-        modules.add("@microsoft/dynamicproto-js", "./node_modules/@microsoft/dynamicproto-js/lib/dist/umd/dynamicproto-js", true);
-
-        // Load Core
-        modules.add("@microsoft/applicationinsights-core-js", "./node_modules/@microsoft/applicationinsights-core-js/browser/applicationinsights-core-js");
-
-        // Load Common
-        modules.add("@microsoft/applicationinsights-common", "./node_modules/@microsoft/applicationinsights-common/browser/applicationinsights-common");
-
-        modules.add("react-native");
-
-        // Load React native device info
-        modules.add("react-native-device-info", "./node_modules/react-native-device-info/lib/commonjs/web/index");
-        
-        var testModule = modules.add("Tests/Unit/src/reactnativeplugin.tests", "./Unit/dist/reactnativeplugin.tests.js")
-        testModule.run = function (tests) {
-            console && console.log("Starting tests");
-            QUnit.start();
-            tests.runTests();
-        };
-
-        modules.run();
-    </script>
-</head>
-
-<body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-    <div id="error-message"></div>
-</body>
-
-</html>
