--- extensions/applicationinsights-react-native/Tests/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "sourceMap": true,
-        "inlineSources": true,
-        "noImplicitAny": false,
-        "module": "amd",
-        "moduleResolution": "Node",
-        "target": "es5",
-        "alwaysStrict": true,
-        "declaration": true,
-        "outFile": "./Selenium/reactnativeplugin.tests.js"
-      },
-      "include": [
-      ]
-}
