--- extensions/applicationinsights-react-native/package.json
@@ -1,70 +1,0 @@
-{
-    "name": "@microsoft/applicationinsights-react-native",
-    "version": "2.5.5",
-    "description": "Microsoft Application Insights React Native Plugin",
-    "main": "dist-esm/index.js",
-    "types": "types/index.d.ts",
-    "sideEffects": false,
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/microsoft/ApplicationInsights-JS/tree/master/extensions/applicationinsights-react-native"
-    },
-    "scripts": {
-        "build": "npm run build:esm && npm run build:package && npm run dtsgen",
-        "build:esm": "grunt reactnative",
-        "build:package": "rollup -c",
-        "rebuild": "npm run build",
-        "test": "grunt reactnativetests",
-        "mintest": "grunt reactnative-mintests",
-        "testx": "npm run build:test && grunt reactnativetests",
-        "lint": "tslint -p tsconfig.json",
-        "dtsgen": "api-extractor run --local && node ../../scripts/dtsgen.js \"Microsoft Application Insights react native plugin\"",
-        "ai-min": "grunt reactnative-min",
-        "ai-restore": "grunt reactnative-restore"
-    },
-    "devDependencies": {
-        "@microsoft/ai-test-framework": "0.0.1",
-        "@microsoft/applicationinsights-rollup-es3": "1.1.3",
-        "@microsoft/api-extractor": "^7.18.1",
-        "grunt": "^1.5.3",
-        "grunt-cli": "^1.4.3",
-        "grunt-contrib-qunit": "^5.0.1",
-        "grunt-contrib-uglify": "^5.0.1",
-        "@nevware21/grunt-ts-plugin": "^0.4.3",
-        "@nevware21/grunt-eslint-ts": "^0.2.2",
-        "@typescript-eslint/eslint-plugin": "^4.28.0",
-        "@typescript-eslint/parser": "^4.28.0",
-        "eslint": "^7.29.0",
-        "eslint-config-standard": "^16.0.3",
-        "eslint-plugin-import": "^2.23.4",
-        "eslint-plugin-node": "^11.1.0",
-        "eslint-plugin-promise": "^5.1.0",
-        "qunit": "^2.11.2",
-        "react": "^17.0.2",
-        "react-native": "^0.68.0",
-        "react-native-device-info": "^5.6.5",
-        "globby": "^11.0.0",
-        "magic-string": "^0.25.7",
-        "@rollup/plugin-commonjs": "^18.0.0",
-        "@rollup/plugin-node-resolve": "^11.2.1",
-        "@rollup/plugin-replace": "^2.3.3",
-        "rollup-plugin-cleanup": "^3.2.1",
-        "rollup-plugin-peer-deps-external": "^2.2.4",
-        "rollup": "^2.32.0",
-        "typescript": "^4.3.4",
-        "tslib": "^2.0.0",
-        "uglify-js": "3.16.0"
-    },
-    "dependencies": {
-        "@microsoft/applicationinsights-common": "2.8.5",
-        "@microsoft/applicationinsights-core-js": "2.8.5",
-        "@microsoft/applicationinsights-shims": "2.0.1",
-        "@microsoft/dynamicproto-js": "^1.1.6"
-    },
-    "peerDependencies": {
-        "tslib": "*",
-        "react-native": "*",
-        "react-native-device-info": "^5.2.1"
-    },
-    "license": "MIT"
-}
