--- extensions/applicationinsights-react-native/src/Interfaces/IReactNativePluginConfig.ts
@@ -1,4 +1,0 @@
-export interface IReactNativePluginConfig {
-    disableDeviceCollection?: boolean;
-    disableExceptionCollection?: boolean;
-}
