--- extensions/applicationinsights-react-native/src/Interfaces/index.ts
@@ -1,4 +1,0 @@
-import { INativeDevice } from "./INativeDevice";
-import { IReactNativePluginConfig } from "./IReactNativePluginConfig";
-
-export { INativeDevice, IReactNativePluginConfig };
