--- extensions/applicationinsights-react-native/src/index.ts
@@ -1,4 +1,0 @@
-import {ReactNativePlugin} from "./ReactNativePlugin";
-import { INativeDevice, IReactNativePluginConfig } from "./Interfaces";
-
-export { ReactNativePlugin, INativeDevice, IReactNativePluginConfig };
