--- extensions/applicationinsights-react-native/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "sourceMap": true,
-    "inlineSources": true,
-    "noImplicitAny": false,
-    "module": "es6",
-    "moduleResolution": "Node",
-    "target": "es5",
-    "alwaysStrict": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "allowSyntheticDefaultImports": true,
-    "importHelpers": true,
-    "noEmitHelpers": true,
-    "forceConsistentCasingInFileNames": true,
-    "declaration": true,
-    "declarationDir": "extensions/applicationinsights-react-native/types",
-    "outDir": "dist-esm",
-    "rootDir": "extensions/applicationinsights-react-native/src"
-  },
-  "include": [
-    "./src/**/*.ts"
-  ],
-  "exclude": [
-    "./node_modules/**"
-  ]
-}
