--- extensions/applicationinsights-react-native/tslint.json
@@ -1,5 +1,0 @@
-{
-    "extends": [
-        "../../tslint-base.json"
-    ]
-}
