--- nuget/Microsoft.ApplicationInsights.JavaScript.Snippet.nuspec
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>Microsoft.ApplicationInsights.JavaScript.Snippet</id>
-    <version>1.0.0</version>
-    <title>Application Insights Snippet for JavaScript Applications</title>
-    <authors>Microsoft</authors>
-    <owners>Microsoft</owners>
-    <language>en-us</language>
-    <licenseUrl>http://go.microsoft.com/fwlink/?LinkID=510709</licenseUrl>
-    <projectUrl>http://go.microsoft.com/fwlink/?LinkId=392727</projectUrl>
-     <icon>images\icon.png</icon>
-    <requireLicenseAcceptance>true</requireLicenseAcceptance>
-    <summary>Application Insights Snippet for JavaScript Applications</summary>
-    <description>Application Insights Snippet for JavaScript Applications, more infor about snipept available in https://github.com/microsoft/ApplicationInsights-JS</description>
-    <tags>Analytics Azure ApplicationInsights Telemetry Monitoring SDK</tags>
-    <copyright>© Microsoft Corporation. All rights reserved.</copyright>
-  </metadata>
-  <files>
-    <file src="..\AISKU\snippet\snippet.min.js" target="content\scripts\ai-snippet.min.js"/>
-    <file src="icon.png" target="images" />
-  </files>
-</package>
